+++ common-mvc/src/main/java/org/todaybook/commonmvc/autoconfig/security/TodayBookSecurityProperties.java
@@ -1,0 +1,8 @@
+package org.todaybook.commonmvc.autoconfig.security;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "todaybook.security")
+public record TodayBookSecurityProperties(Mvc mvc) {
+  public record Mvc(boolean enabled) {}
+}
